Scrape_Master/requirements.txt
litellm
python-dotenv
pydantic
pandas
openpyxl
streamlit-tags
supabase
streamlit
streamlit-tags
crawl4ai
fastapi
uvicorn[standard]
requests
nest-asyncio